Ping-Pong Level 2 Code
Some of the string will be depending on your file name, like the actor line. Make sure to rename it to the file name you are using. The rest of the code can be more or less the same if you want to copy it directly. 
Also, create the sounds and images folder inside your project folder. Put the assets you use inside respective folder.